various changes
1997-09-15
Always use FASTF to position a new tape to a non-zero footage.

1997-09-15
Check error return when searching for the first scan in case
this station has no scans in this schedule.

1997-09-29
Calculate ift_save only for non-S2 continuous tape motion.

1997-09-29
Calculate slewing time only for continuous tape motion.

1997-10-03
For late stop stations, don't force a stop on the last scan
because it is handled separately at the end of the routine.