various changes
2003-05-19
improved programming structure, which makes it easier
for the user to edit the code.
Made gnplt compare 'range' LOs to the actual LO, instead
of sky frequency.

2003-06-02
many changes
allow user to change the rxg file directory
changed the autoscaling options.
changed the meaning of median for fitting.
changed the way gnplt deletes points when zooming/rescaling.
changed 'string range' so old versions of tcl can use it
the 'delete points with bad gain compression' now works better
fixed the columnbreak for source/frequency menus
added a box next to plot with the number of points highlighted,
selected, included and total.
changed the order in which the cascaded menus are displayed
both the fitted and the original tcal curve now have checkbuttons
added an extra decimal to the gain output
made sure MAXLINE is used properly
made the gain curve tools menu options consistent with tcal curve
added stuff to the help menu
fixed bugs in the 'area delete' option
the gain shortcut now computes the gain using the option selected
on the axis menu, not always the first option as before.