update nng and new makefile generator
This updates the third party library nng, and includes a new script to
generate the Makefile. NNG requires CMake version >= 3.12 to build, but
this is not available in Debian 9 or earlier, so we make an unofficial
Makefile with the configuration for Debian 9 baked in. CMake >= 3.12 is
required to generate the Makefile but not after that. Static builds of
CMake are available.

This is yet to be tested on Debian 7 (FSL9) or Debian 5 (FSL8).