fsserver: wait for old fs window to finish shutdow
If a new FS session is immediately requested while the old window is
still running in its "shutdown" state, we must wait for this to
complete to prevent two sockets trying to hold the same address.

This causes slight delay for imediate restarts, but this can't be
avoided without allowing multiple fs windows.