fsclient: ignore prompt in no-X11 mode
If FS client is in no-X11 mode, it currenly still creates a `fs.prompt`
when instructed by the server. This change removes that behaviour,
though it may cause an issue if no other clients exist to dismiss the
prompt, see issue #49. If this is a problem for anyones usecase we will
need a new feature here.

Closes #49