Fix compiling nng-1.3.0 on lenny
This update brings the baked-in NNG Makefile in line with Debian 5 (lenny).

The necessary minor changes were derived using CMake 3.6.3 (which is able
to process the nng-1.3.0 source apart from the "CheckAtomicLib" directive
which apparently is not relevant on x86) on a FSL8 platform using the
methodology of the gen_nng.make Makefile generator.  Required changes are:

	o Substitute platform/posix/posix_pollq_poll.c for posix_pollq_epoll.c
	  (as was previously done in the nng-1.0.1 Makefile).
	o Eliminate the CPPFLAGS defines -DNNG_HAVE_EPOLL_CREATE1=1 (reflecting
          the simpler content of sys/epoll.h) and -DNNG_HAVE_EVENTFD (due to
          the lack of a sys/eventd.h header) to reflect Debian 5 capabilities.