Log DBBC3 multicast based on time interval
Previously, multicast message receptions were used. That was fine
for a one second logging interval and for longer intervals if there
weren't many times-outs. However there can be suppressed messages if
the DBBC3 is busy, This could lead to a long gap in logged data if the
interval is long and the message at the logging interval isn't received.
Now the next message received after the interval has expired is logged.