Revert "Remove trying to disable signals"
This reverts commit 10047490899497f18d34f4045e74a92cc2f56e4f.

This is not technically needed. However, if it isn't present and the
display server is not in use, a Control-C in the window where the FS
was started will leave the FS in a partially terminated state because
of an extraneous call to sig_ignore() in ddout.c. The extraneous call
will be removed, but it is marginally more beneficial to let the FS
shutdown more gracefully anyway.

With Control-C ignored in general it will only kill xterms, typically
only for oprin, opened on FS at start-up (there could also station
programs that don't use setup_ids() that will be killed). That will
allow the rest of the FS to shutdown gracefully. This also gives the
FS without the display server the lowest profile possible to avoid
being accidentally killed (not very low though).

Extraneous (redundant) calls to sig_ignore() in ddout.c and oprin.c
will be removed in the next commit.