Remove limited use of prototypes
Removed clib.h/poclb.h/rtelb.h and their use.

They created an incompatibility for the float in fltstr() for calls
that did not have the prototype. With the prototype the float value is
a float; without it, it is a double. This incremental prototyping
approach should be reconsidered.

Unfortunately this change causes compile warning for dbtcn.