Upgrade of IBCON to handle general TCP/IP network devices
This version allows IBCON to be used for communication with any
network device using TCP/IP sockets. It has been in use at Onsala Space
Observatory since 2013.

The documentation is in misc/general_ibcon.pdf

The 'rdtib.f' file in 'boss' should be replaced with this version
which allows longer lines in 'ibad.ctl', when the OSO version of
'ibcon' is used.