Fix DBBC/3 tpicd Tsys being too low by Tcal/2
The CALon data was not being included the Tsys level even though it
represents the level half of the time. Now the CALon and CALoff
values are averaged.

The DBBC3 support will not use this direct communication approach
after the multicast support is merged but the code is correct now in
case it every gets reused. In the testing for this, it looked like the
speed of the direct communication has been improved enough that it
would be a viable option now, but the multicast is superior. To
re-enable the direct communication approach, the leading DBBC3 clause
in tpicd that diverts the control to dbtcn for multicast would need to
be removed. However, the Tsys in the new DBBC3 monitor display would
still come from multicast unless more work is done on that.

Closes #157