Improve DBBC3 multicast time-out suppression
All time-outs are reported when data_valid is on. All time-outs when
data_valid is off and a DBBC3 command may have interfered with
multicast generation are suppressed.

Two conditions are needed for the data_valid off case to know if a
DBBC3 command could have caused the time-out:

1. Has the count of DBBC3 commands (counting both the start and ends
   of commands) changed since before the previous select()?

2. Is a DBBC3 command active when the time-out is detected?

Case 1 is needed to handle if a command was executed that interfered
with multicast before the current select(). This may cause suppression
of some non-DBBC3 command caused time-outs, but only if they occurred
within, at most, two time-out periods (2.5 seconds total) of the last
DBBC3 command. Both the start and ends of commands (as opposed to just
the starts) have be counted in case the command took so long to
execute that the start was before the previous select().

Case 2 is needed in case a DBBC3 command started before the previous
select() and still hasn't finished when the time-out occurs.