Set data_send=off before changing RDBE time
For each transaction that changes the time or sync, the state of data
sending for that RDBE is checked. If it is 'on', it is turned 'off'
and there is a wait of slightly more than one second to make sure
there at least one PPS has passed before the requested operation is
performed. If sending is already 'off', there is no wait, By checking
each transaction, fmset does its best to handle sending being turned
'on' by the operator while fmset is already running. This is for
safety. Generally, there should only be one wait per RDBE per fmset
run.

When fmset is exited, it will turn sending 'on', per RDBE, if it had
turned sending 'off, making sure at least one PPS has passed since the
last operation before turning it 'on'. Effectively sending will be
turned 'on' at exit if fmset ever detected it being 'on' when it
performed an operation in the current run, regardless if the operator
changed sending to 'off' while fmset was running.

If fmset crashes, preventing it from re-enabling data sending, it will
have to be turned 'on' by the operator, but the next RDBE setup should
do that in any event.