Add control file for erchk
The syntax is explained in /usr2/fs/st.default/control/erchk.ctl.
That default/example implements the previous behavior except that 'sp'
errors are now shown instead of being suppressed. With that
configuration, all errors are displayed. Individual numeric errors can
now be displayed differently. Additional display attributes are now
accessible.

The non-working elimination of duplicate sequential error messages was
removed. That was broken (if it ever worked) when the display server
was added. It is not consistent with the new default of reporting all
errors, so it was not fixed.

Increased the accuracy of the error code matching, so some codes aren't
incorrectly matched if they appear in the error text.

Closes #174