Protect setcl from DBBC3 v124 and unsynced boards
This includes trapping a VDIF seconds of 1. This prevents a bogus time
from being accepted for the case of v124 being loaded (which returns 1
in the multicast for an unsynced board), but v125 or higher being
claimed in the FS control files. If 1 is ever encountered for v125 or
higher, the retry will get a larger value.