Invalidate pair match if the string doesn't agree
This prevents a match on the second command of a pair if the first
command of that pair with a different string has occurred since the
original first command with the right string.

This prevents matching the wrong second command, if the one between
two first commands with different strings is missing for some reason,
like it timed-out.

Closes #182