Add RDBE pcal plotting
The default for RDBE pcal is to plot just the first tone encountered
for each IF. The set of tones plotted can be overridden with the '-d'
and '-D' options.

The -j and -k options (Tsys normalization) are not supported for RDBE
pcal yet.

The -v (version) option was moved to -V. Now -v is phase differences
within VGOS bands.