Add -h -w -s options, TRAP fix, tail restructured
Added options including help, wait, and scroll-back.

Don't set TRAP for EXIT until expecting a 'tail' process to exist;
this avoids spurious errors if control-C is used during a wait.

The use of tail was changed to provide a limited scroll-back function
that may help pick up some missing lines when there is a log change if
the FS isn't generating a lot of output at that point.