Add 'giza' adjustments for variances from 'pgplot'
The 'giza' adjustments are reducing the line width and using a larger
open circle for bad points. The '-Z hex' option can be used to
selectively remove the adjustments if they cause problems or are not
wanted. If 'giza' is not detected, the '-Z hex' option can be used to
force the adjustments.