Workaround f2c limitation concatenating strings
Apparently f2c (FC=fort77) cannot handle making a temporary string
from a concatenation that includes a character*(*) adjustable length
argument. f95 handles this okay, but having different code for
different compilers seemed too complicated to be worth it. Instead, a
local declared string variable of the maximum expected length (set by
the MAX_SKD parameter) is used. If its length is too small for the
string passed, either (i) a warning is issued if the needed non-blank
part of the string is small enough to handle, or (ii) an error is
issued if it is not. If either occurs, the code, probably in the
caller, needs to be corrected. All paths currently work.