Improve rdbe echo output for long responses
The maximum possible response length is 8192 bytes, per Russ. For a
long response to be returned to the caller, it has to be truncated to
the maximum class message size (currently 2048), including the
trailing NULL. An error is reported if truncation is necessary. The
echo (debug) output is used to show the full response.